backend/requirements.txt
Flask==2.3.3
PyJWT==2.8.0
bcrypt==4.0.1
mysql-connector-python==8.1.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
flasgger==0.9.7.1
flask-cors==4.0.0